bugfix: renaming classes
